#689 Moving test to right spot and making it Locale independent.
