+++ processor/src/test/java/org/mapstruct/ap/test/builtin/jodatime/bean/LocalDateBean.java
@@ -16,7 +16,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.mapstruct.ap.test.builtin.bean.jodatime.bean;
+package org.mapstruct.ap.test.builtin.jodatime.bean;
 
 import org.joda.time.LocalDate;
