+++ processor/src/test/java/org/mapstruct/ap/test/builtin/jodatime/mapper/DateTimeToXmlGregorianCalendar.java
@@ -16,12 +16,12 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.mapstruct.ap.test.builtin.bean.jodatime.mapper;
+package org.mapstruct.ap.test.builtin.jodatime.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-import org.mapstruct.ap.test.builtin.bean.jodatime.bean.LocalDateTimeBean;
-import org.mapstruct.ap.test.builtin.bean.jodatime.bean.XmlGregorianCalendarBean;
+import org.mapstruct.ap.test.builtin.jodatime.bean.DateTimeBean;
+import org.mapstruct.ap.test.builtin.jodatime.bean.XmlGregorianCalendarBean;
 import org.mapstruct.factory.Mappers;
 
 /**
@@ -29,10 +29,10 @@ import org.mapstruct.factory.Mappers;
  * @author Sjaak Derksen
  */
 @Mapper
-public interface XmlGregorianCalendarToLocalDateTime {
+public interface DateTimeToXmlGregorianCalendar {
 
-    XmlGregorianCalendarToLocalDateTime INSTANCE = Mappers.getMapper( XmlGregorianCalendarToLocalDateTime.class );
+    DateTimeToXmlGregorianCalendar INSTANCE = Mappers.getMapper( DateTimeToXmlGregorianCalendar.class );
 
-    @Mapping( target = "localDateTime", source = "xMLGregorianCalendar" )
-    LocalDateTimeBean toDateTimeBean( XmlGregorianCalendarBean in );
+    @Mapping( target = "xMLGregorianCalendar", source = "dateTime")
+    XmlGregorianCalendarBean toXmlGregorianCalendarBean( DateTimeBean in );
 }
